+++ src/utils/extendedOnChangeBoolean.jsx
@@ -4,6 +4,7 @@ const extendedOnChangeBoolean = (onChangeFn, event) => {
     if (_.isFunction(onChangeFn)) {
         onChangeFn({
             event,
+            name: event.target.name,
             value: event.target.checked,
             rawValue: event.target.value,
         });