+++ ui-test/tests/Inputs.jsx
@@ -9,7 +9,6 @@ import {
     CardContent,
     Radio,
     RadioGroup,
-    SelectBox,
     Switch,
     TextField,
     DateField,
@@ -19,10 +18,6 @@ import {
 const TestInputs = React.createClass({
     getInitialState() {
         return {
-            selectBox1: {label: 'labelz', value: 'valuez'},
-            selectBox2: 8,
-            selectBox3: '',
-            selectBox4: '',
             switches: [false, true, undefined, undefined, true, false],
             textInput: ['5'],
             selectedRadio: '',
@@ -43,61 +38,16 @@ const TestInputs = React.createClass({
             ],
         };
     },
-
-    selectBox1OnChange(value) {
-        console.log('SelectBox onChange: ', value);
-        this.setState({selectBox1: value});
-    },
-    selectBox2OnChange(value) {
-        console.log('SelectBox onChange: ', value);
-        this.setState({selectBox2: value});
-    },
-    selectBox3OnChange(value) {
-        console.log('SelectBox onChange: ', value);
-        this.setState({selectBox3: value});
-    },
-    selectBox4OnChange(value) {
-        console.log('SelectBox onChange: ', value);
-        this.setState({selectBox4: value});
-    },
-
-    updateSwitch(index, {value}) {
-        const switches = _.clone(this.state.switches);
-        switches[index] = value;
-        this.setState({
-            switches,
-        });
-    },
-
-    updateRadio({value}) {
-        this.setState({
-            selectedRadio: value,
-        });
-    },
-
-    updateTextInput(index, {value}) {
-        const textInput = _.clone(this.state.textInput);
-        textInput[index] = value;
-        this.setState({
-            textInput,
-        });
-    },
-
-    updateDateInput(index, {rawValue, value, isValid}) {
+    updateValue({name, rawValue, value, isValid}) {
         console.log(
-            'New Date is:',
-            value,
+            `Changing value of ${name} to: ${value}`,
             `(${isValid ? 'valid' : 'invalid'})`,
-            'Raw:',
-            rawValue
+            ` | Raw: ${rawValue}`
         );
-        const dateInput = this.state.dateInput;
-        dateInput[index] = value;
-        this.setState({
-            dateInput,
-        });
+        const currentState = this.state;
+        _.set(currentState, name, value);
+        this.setState(currentState);
     },
-
     render() {
         return (
             <Card>
@@ -105,50 +55,56 @@ const TestInputs = React.createClass({
                 <CardContent>
                     <Switch
                         id="test_id_666"
+                        name="switches[0]"
                         ripple
                         checked={this.state.switches[0]}
-                        onChange={this.updateSwitch.bind(null, 0)}
+                        onChange={this.updateValue}
                     />
                     <Switch
+                        name="switches[1]"
                         checked={this.state.switches[1]}
-                        onChange={this.updateSwitch.bind(null, 1)}>
+                        onChange={this.updateValue}>
                         Switch with Ripple
                     </Switch>
                     <hr />
                     <Checkbox
+                        name="switches[2]"
                         id="test_id_667"
                         ripple
                         checked={this.state.switches[2]}
-                        onChange={this.updateSwitch.bind(null, 2)}
+                        onChange={this.updateValue}
                     />
                     <Checkbox
+                        name="switches[3]"
                         label="Checkbox 1 Text"
                         checked={this.state.switches[3]}
-                        onChange={this.updateSwitch.bind(null, 3)}
+                        onChange={this.updateValue}
                     />
                     <Checkbox
+                        name="switches[4]"
                         disabled
                         checked={this.state.switches[4]}
-                        onChange={this.updateSwitch.bind(null, 4)}>
+                        onChange={this.updateValue}>
                         Checkbox 2 Text
                     </Checkbox>
                     <Checkbox
+                        name="switches[5]"
                         checked={this.state.switches[5]}
-                        onChange={this.updateSwitch.bind(null, 5)}>
+                        onChange={this.updateValue}>
                         <div className="test">Checkbox 3 Text</div>
                     </Checkbox>
                     <hr />
                     <RadioGroup
-                        name="testradio1"
-                        onChange={this.updateRadio}
+                        name="selectedRadio"
+                        onChange={this.updateValue}
                         value={this.state.selectedRadio}>
                         <Radio value="one" />
                         <Radio value="two" label="Radio 2 Text" />
                     </RadioGroup>
                     <RadioGroup
                         childContainer="div"
-                        name="testradio2"
-                        onChange={this.updateRadio}
+                        name="selectedRadio"
+                        onChange={this.updateValue}
                         value={this.state.selectedRadio}>
                         <Radio disabled value="three">
                             Radio 3 Text
@@ -190,108 +146,34 @@ const TestInputs = React.createClass({
                         </Radio>
                     </RadioGroup>
                     <hr />
-                    <h5>SelectBox with objects</h5>
-                    <SelectBox
-                        placeholder={'Value deleted'}
-                        options={[
-                            {
-                                label: 'labelz',
-                                value: 'valuez',
-                            },
-                            {
-                                label: 'label1',
-                                value: 'value1',
-                            },
-                            {
-                                label: 'label2',
-                                value: 'value2',
-                            },
-                        ]}
-                        value={this.state.selectBox1}
-                        onChange={this.selectBox1OnChange}
-                        autofocus
-                    />
-                    <h5>SelectBox with mixed strings and numbers</h5>
-                    <SelectBox
-                        placeholder={'No Value'}
-                        options={['label1', 3, 8]}
-                        value={this.state.selectBox2}
-                        onChange={this.selectBox2OnChange}
-                    />
-                    <SelectBox
-                        placeholder={'Open option to top'}
-                        options={['label1', 3, 8]}
-                        value={this.state.selectBox2}
-                        onChange={this.selectBox2OnChange}
-                        optionsOnTop
-                    />
-                    <h5>
-                        Multiple selections with option to create new entries
-                    </h5>
-                    <SelectBox
-                        placeholder={'Multiselect with create'}
-                        options={[
-                            'label1',
-                            3,
-                            'Andelsselskab med begrænset ansvar, Denmark',
-                            'Akcionersko Društvo / Акционерско друштво, Republic of Macedonia',
-                            8,
-                        ]}
-                        value={this.state.selectBox3}
-                        onChange={this.selectBox3OnChange}
-                        multi
-                        clearable={false}
-                        creatable
-                        promptTextCreator={newLabel =>
-                            `New freaking Stuff called: ${newLabel}`
-                        }
-                    />
-                    <h5>Async Selectbox</h5>
-                    <SelectBox
-                        placeholder={'Multiselect with async'}
-                        value={this.state.selectBox4}
-                        onChange={this.selectBox4OnChange}
-                        loadOptions={function(input, callback) {
-                            setTimeout(() => {
-                                callback(null, {
-                                    options: [
-                                        {value: 'one', label: 'One'},
-                                        {value: 'two', label: 'Two'},
-                                    ],
-                                    // CAREFUL! Only set this to true when there are no more options,
-                                    // or more specific queries will not be sent to the server.
-                                    complete: true,
-                                });
-                            }, 500);
-                        }}
-                        multi
-                        async
-                    />
-                    <hr />
                     <h5>Textfields</h5>
                     <TextField
+                        name="textInput[0]"
                         value={this.state.textInput[0]}
                         label="Test Input"
-                        onChange={this.updateTextInput.bind(null, 0)}
+                        onChange={this.updateValue}
                     />
                     <TextField
+                        name="textInput[0]"
                         value={this.state.textInput[0]}
                         label="Test Input (not full width)"
-                        onChange={this.updateTextInput.bind(null, 0)}
+                        onChange={this.updateValue}
                         stretch={false}
                     />
                     <TextField
                         multiline
+                        name="textInput[1]"
                         label="Multiline input"
                         className="mdl-textfield--full-width"
                         value={this.state.textInput[1]}
-                        onChange={this.updateTextInput.bind(null, 1)}
+                        onChange={this.updateValue}
                     />
                     <TextField
+                        name="textInput[0]"
                         value={this.state.textInput[0]}
                         label="Test Input"
                         error="Something went wrong, so this error is shown."
-                        onChange={this.updateTextInput.bind(null, 0)}
+                        onChange={this.updateValue}
                     />
                     <h5>Datefields</h5>
                     date only fields (8th, February 2017)<br />
@@ -300,56 +182,66 @@ const TestInputs = React.createClass({
                         stretch={false}
                         error="This error is shown."
                         value={this.state.dateInput[0]}
-                        onChange={this.updateDateInput.bind(null, 0)}
+                        name="dateInput[0]"
+                        onChange={this.updateValue}
                         className="my-uitest-class"
                     />
                     <DateField
                         placeholder="Date placeholder"
                         stretch={false}
                         value={this.state.dateInput[1]}
-                        onChange={this.updateDateInput.bind(null, 1)}
+                        name="dateInput[1]"
+                        onChange={this.updateValue}
                     />
                     <DateField
                         value={this.state.dateInput[2]}
-                        onChange={this.updateDateInput.bind(null, 2)}
+                        name="dateInput[2]"
+                        onChange={this.updateValue}
                     />
                     <DateField
                         value={this.state.dateInput[3]}
-                        onChange={this.updateDateInput.bind(null, 3)}
+                        name="dateInput[3]"
+                        onChange={this.updateValue}
                     />
                     <DateField
                         value={this.state.dateInput[4]}
-                        onChange={this.updateDateInput.bind(null, 4)}
+                        name="dateInput[4]"
+                        onChange={this.updateValue}
                         closeOnSelect
                         placeholder="picker disappear on select"
                     />
                     time only fields (14:23)<br />
                     <DateTimefield
                         value={this.state.dateInput[5]}
-                        onChange={this.updateDateInput.bind(null, 5)}
+                        name="dateInput[5]"
+                        onChange={this.updateValue}
                         dateFormat={false}
                     />
                     <DateTimefield
                         value={this.state.dateInput[6]}
-                        onChange={this.updateDateInput.bind(null, 6)}
+                        name="dateInput[6]"
+                        onChange={this.updateValue}
                         dateFormat={false}
                     />
                     <DateTimefield
                         value={this.state.dateInput[7]}
-                        onChange={this.updateDateInput.bind(null, 7)}
+                        name="dateInput[7]"
+                        onChange={this.updateValue}
                         dateFormat={false}
                     />
                     date and time
                     <DateTimefield
                         value={this.state.dateInput[8]}
-                        onChange={this.updateDateInput.bind(null, 8)}
+                        name="dateInput[8]"
+                        onChange={this.updateValue}
                     />
                     no manual input
                     <br />
                     Selected date: {this.state.dateInput[9].format()}
                     <DateTimefield
                         value={this.state.dateInput[9]}
-                        onChange={this.updateDateInput.bind(null, 9)}
+                        name="dateInput[9]"
+                        onChange={this.updateValue}
                         input={false}
                     />
                 </CardContent>